requirements.txt
langdetect==1.0.9
pytorch_lightning==2.0.0
torchmetrics==0.11.4
torch==2.0.0
scikit-learn==1.2.2
transformers[torch]==4.34.1
hydra-core==1.3.2
omegaconf==2.3.0